Created virtual environment installation script.
scripts/install_venv.sh should install all dependencies for the
project to a virtual environment created by the script in a folder
beside the repository. Working on DICE. Takes up about 3.0GB total and
around 20 minutes to compile, so be careful to check your quota usage
before you try this.