Basic input recognition from any gamepad.
First stages of the gamepad controller for the robot. Allows selection
of a joypad from all devices currently connected to the machine. Prints
output when a button is pressed or an axis changes position.

Current intention is to only support XBOX 360 controllers, but it should
be easy to extend to any gamepad.